fix: client connected InvokeOnClientConnectedCallback with scene management disabled (#1123)
* fix

Fix for the client connected message not being invoked on the client if scene management is disabled.
This includes an integration test, MultiClientConnectionApproval.ClientConnectedCallbackTest,  to verify client connected is being invoked when scene management is both enabled and disabled.

* fix

This is a fix to reduce the amount of console spew due to network buffer pools exceeding 1024.  This reduces unit test processing time and console log sizes, while also providing a more realistic buffer pool size.

* test

The unit test.

* refactor

Removing this minor LF as the file has nothing to do with this PR.

* style

removing 2 LF that were not intended

* refactor style

Some straggler style requests from PR-1080 that are harmless and piggy-backing onto this PR.

* refactor

Reverting the NetworkBufferPool changes.
This will make MessageHandlerReceivedMessageServerClient fail because of the warning message that we have exceeded 1024 buffers.