refactor fix
This is just an update on the connection approval manual test to verify that users receive the appropriate messages when a client is approved or not approved.
Server will know if the client is approved or not based on how it processed the NetworkConfig.ConnectionData sent to it and clients will just be notified they were disconnected.