fix
When a client or server have mismatched connection approval NetworkManager settings this fix will make sure to disconnect the client as well as it will properly detect the mismatched NetworkConfigs and log a message to the console on the server/host so users know why the client was disconnected.

Made an adjustment to the connection approval component so this scenario can be tested.