refactor - update
Added a SceneAwareNetworkObject that will automatically synchronize clients when a NetworkObject is moved into a different scene.  This should provide users with one example of how they can keep dynamically spawned NetworkObjects in synch with the proper scene it should be located within when using additive scenes.

Adjusted the parenting code to leverage from this class in order to better simplify the scene synchronization process.

Fixed an issue with randomly dropped RPCs since I was parenting within the collision handler.  Now the parenting happens in the update to avoid missing RPCs during FixedUpdate (this is a temporary fix until this issue with dropped RPCs is resolved).