refactor
This update allows one to add an additional start initialization action to be invoked when just after the server-host is started and after each client is started for scenarios where you need to access sub-systems that are only initialized when the NetworkManager is started (client, host, or server).

Made minor modifications to the BaseMultiInstanceTest to reflect the above changes.