test: make test runner scene ignored by default for BaseMultiInstanceTest (#1154)
* refactor

Specific to testing, this adds the ability to ignore the warning message when scene validation fails so we don't spam log.

* refactor

If not already set, making the test runner scene a default scene for clients to not try to synchronize.  This also will register the test runner scene so that other associated warning messages will not log-spam.

* refactor

This update allows one to add an additional start initialization action to be invoked when just after the server-host is started and after each client is started for scenarios where you need to access sub-systems that are only initialized when the NetworkManager is started (client, host, or server).

Made minor modifications to the BaseMultiInstanceTest to reflect the above changes.

* style

included 1 whitespace

* style

updated comment and removed a LF

* style

updated comment.

* refactor

updated comment.
using startInitializationCallback?.Invoke(server); as opposed to checking for null.