Only track one metric for scene sync and do not report scene name (#1159)
Previously we were tracking all scenes that were included in the message when an S2C_Sync message was sent or received. Now we have a better understanding of this scene event type, this feels like the wrong level of detail for this metric so we are now just reporting the scene event type and not the specific scenes.

This will also make this metric consistent for C2S_SyncComplete where the scene names is not a core part of the message (the important thing is the event type itself and the number of network objects synced, not the scenes themselves).

Similarly, S2C_Resync will not report each scene name, to keep consistent with the rest of the scene flow.