refactor
This allows for two different types of client synchronization modes (ClientSynchronizationMode):
LoadSceneMode.Single: All currently loaded scenes on the client will be unloaded and the server's currently active scene will be loaded in single mode on the client unless it was already loaded.

LoadSceneMode.Additive:  All currently loaded scenes are left as they are and any newly loaded scenes will be loaded additively.  Uses need to determine which scenes are valid to load via the VerifySceneBeforeLoading method.