fix
This fixes Github issue 1102.
The issue was that the server was running faster than the disk I/O for scene loading and so the SceneEventData single instance was being written over before the client was done loading.
To avoid this issue and any future potential issues I created a SceneEventData pool and use a unique SceneEventData instance for each scene event.  Currently set the default pool size to 10 but made it so it can be adjusted if needed through the public facing method NetworkSceneManager.SetSceneEventDataPoolSize.