fix and refactor
This updates the most recent NetworkTransform changes to be able to handle NetworkObject pools.

The adjustments made are reflective to the rest of the test project changes.
Added the ability to toggle interpolation on and off within the NetworkPrefabHandlerObjectPool classes in order to see what interpolation looks like when it is on and off real time.