fix: Flush the UnityTransport send queue on shutdown (#1234)
* fix: Flush the UnityTransport send queue on shutdown

When NetworkManager.Shutdown is called, all clients are disconnected and
the transport is immediately shut down. For UnityTransport, that means
disconnect messages are not actually sent to the remote peers, because
nothing goes out on the wire until an internal update of the driver is
scheduled (which doesn't happen between the disconnect calls and the
transport's shutdown).

This is fixed by flushing the internal send queue when shutting down the
transport. This will send out any disconnect messages that might still
be pending. The downside is that UnityTransport.Shutdown might take a
bit longer to execute, since it will need to perform network activity.
But when you're shutting down, you're presumably not in a situation
where performance is critical, so this is probably acceptable.

Co-authored-by: Andrew Spiering <andrews@unity3d.com>