test
Added integration test that validates the new delegate events are triggering at the right times with the right information.
Had to add a BaseMultiInstanceTests.m_BypassStartAndWaitForClients bool value that allows you to bypass starting the host and clients as well as the waiting for the clients to connect to test for the client synchronization notifications (i.e. happens during the client connection period of time which already passes by the time the actual unit test starts).