feat: change the way NetworkAnimator works (its more correct) which reduces the amount of network traffic
fix: standards

more optimization.

feat: Client networkrigidbody (#1202)

What this does is enabled the existing functionality of NetworkRigidbody to make it work with client authority and hook it up with the incoming CanCommit property of NetworkTransform.

fix: improved scene event type names and new events with associated delegate declarations (#1263)

* fix

Renamed the SceneEventTypes to match standards and for better clarity.
Adjusted all places that referenced the previous SceneEventTypes with the new SceneEventTypes.