refactor and fix
Changed when scenes are unloaded to occur once the scene loaded is no longer needed to reduce scene overhead.  Keeping the scenes list to be tested.
SmokTestOrchestrator is constructed during OneTimeSetup now.

Fixed a bug in the SmokTestOrchestrator that would only allow one state to be invoked.

Migrated all logging of what phase a SmokeTestState is in to the SmokeTestState itself.