Refactor and Fix
Refactored how scenes are grouped in order to determine what might need to do scene transitions and what only has 1 scene associated with it (this might go away in the future).

Re-organized a few things to create the SceneAwareSmokeTestState class that includes common scene loading and unloading methods and checks.

Fixed an issue in the NetworkPrefabPool class where it was throwing an exception (seemingly depending upon how many clients were connected) and that would cause an issue with the client's synchronization process.  Just checking for IsSpawned within DeregisterCustomPrefabHandler fixes this issue.