fix
Issue was most likely the NetworkPrefabPoolAdditive class was cleaning itself up during OnDestroy which was causing issues.  Changed it to OnNetworkDespawn and it seemed to fix the issue (previous commit claimed it was 2021.2.0b16 which I no longer think this is the case...but it was indeed when it was cleaning up after itself).

Re-applied the SampleScene UnityTransport adjustment to make it not default to relay.