test
adjusted scene transitioning test to switch scenes every 30 seconds.
increased the number of objects to spawn for each of the 3 scenes to 10 per second.
Running a host and 1 client seems to work with some errors after scene transitioning:
Error sending the message: Unable to allocate packet due to buffer overflow.