fix: UTP adapter tests failing on consoles
A lot of tests were waiting on the Connect event on the server side as a
proxy of the connection being fully established. But that's not correct.
A Connect event on the server does not mean that the UTP Connection
Accept message has been received and processed by the client.

This lead to some timing issues where we'd try to e.g. send data on the
client side before the connection is fully established.