fix: Sends failing on UTP adapter when queue is full (#1333)
* fix: Sends failing on UTP adapter when queue is full

Or close to full. There were two issues here.

First, the payload capacity of the fragmentation pipeline didn't account
for the entire size of the packet. Unfortunately, payload capacity in
the context of the fragmentation pipeline is a bit of a misnomer since
it also includes headers and other such overhead. We were setting it to
the size of a full send queue, and so when you added UTP headers, we
ended up with a packet too large for the pipeline. Bumping up the value
a bit solves this issue.

Second, when sending a single message that was within 4 bytes of the
send queue capacity, we'd silently fail to send it. The reason was we
didn't account for the extra overhead taken by the data length in the
send queue (which requires 4 bytes). Thus the send code would think the
message could fit in the queue when it couldn't. Fixing this simply
required accounting for the 4 extra bytes in the send logic.

Two tests were also added to test sends when the send queue is full
(either from one big send or from multiple smaller ones in a row). Both
tests fail without the fixes in the adapter.

* Cleaner fix for the send queue being full scenario

This one depends on a fix in UTP that's still being reviewed, though.

* Fix possible re-ordering issue for large messages

* Reword the warning when sending immediately

Events can't be larger than m_SendQueueBatchSize. UTP will drop them and
log an error in the current configuration. Reword the warning so that we
don't imply that it's been sent, only that we're trying to send it. If
it fails, there will be an error in the log following it, making it
clear to the user what happened.

* Update UTP dependency to 1.0.0-pre.7

Co-authored-by: Andrew Spiering <andrews@unity3d.com>