fix and refactor
This fixes several areas in the unit/integration tests that were not destroying the actual GameObject but only destroying a component which would leave behind artifact GameObjects.

This migrates the IntegrationTestSceneHandler into its own file, removes dependencies, and integrates it into the MultiInstanceHelper class since this is used by integration/unit tests as a stand alone and is used by the BaseMultiInstanceTest class.

Additional updates to some tests were required due to some of the changes.