fix
Increasing the server side tick wait to be greater than 2 ticks in order to properly calculate the local tick on client side due to how the client applies buffering to the received time.

Removed the MultiInstanceHelpers.Destroy(); from the TestClientTimeInitializationOnConnect method since it is being called in the Teardown method.

Applying isHost parameter to MultiInstanceHelpers.Start as opposed to always being run as a server.