fix: In-scene NetworkObjects disabled but never re-enabled after scene transition (#1354)
* fix

This is the fix for the in-scene NetworkObjects moved into DDOL and the after a scene transition they become disabled but never re-enabled.

* test

Had to make some private methods internal for testing purposes.
Adding test to validate the fix.

* style

white space fixes.

* style

Adding comment above the primary line of code that is the fix for this PR

* style

removing unused namespace.

* style

removing several LFs

* fix and test

This now tracks the state of NetworkObjects when a full scene transition takes place.

This includes a refactored test for NetworkObjects (dynamically spawned and in-scene placed) that persist a full scene transition.

* test update

This makes sure that DestroyWithScene is set to false for all InSceneNetworkObjectState tests.

* style

Removing whitespace.

* updated changelog

Added entries for the fixes applied in this PR.

* refactor

After peer review during PR Triage we all decided to not change or track a NetworkObject's parent GameObject's state and to not disable during full scene transition.

Removed all active state tracking dictionaries and references as well as removed disabling DDOL migrated GameObjects from the runtime NGO code base.

Adjusted the DDOL test to check the state is maintained throughout the scene transition.

The DDOL test only required a minor

* updated changelog

Updating the changelog to reflect most recent updates.