fix
During Client side synchronization, ValidateSceneBeforeLoading should not exit the synchronization process but should ignore the scene in the SceneEventData that the client is being told to load.  This just sets the shouldPassThrough value based on the response from the ValidateSceneBeforeLoading (which will ignore the scene loading process but will still notify that the scene event is being processed).