fix: OnNetworkDespawn not called during shutdown (#1390)
* fix

Renaming and refactoring DestroyNonSceneObjects to be DespawnNetworkObjects as this is really what it is doing.  The destroy portion is either up to the handler, if an in-scene NetworkObject it lets the unloading of the scene destroy it, and the rest are always destroyed.

* fix

updates to the changes in NetworkSpawnManager

* fix

Adding a "ShutdownInProgress" flag to be able to easily check if we are shutting down.

Adding a check for the NetworkManager.LocalClient since we are despawning in the shutdown method.

Ignoring the de-parenting of NetworkObjects when shutting down (this will have to be discussed a bit).

* fix

Test fix:  I think this was just an unexposed bug in the NetworkRigidBody2D test, but will verify this before making this final.

Fixed some additional minor issues like checking for IsSpawned as opposed to NetworkManager.IsListening and IsSpawned as opposed to NetworkObject.IsSpawned.

* fix

Avoiding spam when despawning and shutting down.
Renaming DespawnNetworkObjects to DespawnAndDestroyNetworkObjects.
If a NetworkObject is not spawned it gets destroyed.

* test fix

Fixing issue with editor tests and the NetworkManager.Singleton instance.

* fix

Adjusting the logic a bit on getting the NetworkObject associated with a NetworkBehaviour.

* revert and fix

After discussing this with Luke, it turned out there was a bug in the NetworkRigidBody2D where it was resetting the isKinematic to true.  This reverts the changes made to the NetworkRigidbody2DDynamicTest and fixes the bug with NetworkRigidBody2D.

* test

Tests that validate OnNetworkDespawn is invoked during shutdown.
This tests server, host, and client relative instances.

* update changelog

update the changelog with the fixed information

* fix

fixing issue with test and recent changes for when the internal shutdown method is invoked.

* refactor

ShutdownInProgress now returns m_ShuttingDown.

Setting m_ShuttingDown to false at the end of ShutdownInternal in case the NetworkManager is going to be re-used.

* style

line endings

* Revert "style"

This reverts commit e2ca11a63bd7b10a88a319e0a838c072b0e6c578.

* fix

fixing timing related issue for NetworkObjectOnNetworkDespawnTests.