fix: Address packet overflow issues in the UTP adapter (1.0.0) (#1404)
* Remove maximum packet size setting

It doesn't mean anything now. Anything sent on fragmented pipelines will
be broken up at 1400 bytes boundaries, and everything sent on a
non-fragmented pipeline is limited to a maximum size of 1400 bytes
(including overhead from the underlying protocols). Allowing to set that
limit can only lead to confusion among users.

* Send all reliable data on a fragmented pipeline

This is not the cleanest of solution, but until we figure out something
better in UTP, this can at least unblock Boss Room.

The core issue being fixed here is that we'd allow sending payloads up
to the MTU in size on a non-fragmented pipeline. But that's incorrect
since UTP must add its own overhead to the payload. If sending a payload
close to the MTU in size, the overhead would bring the total packet size
over the limit of the MTU.

Since UTP lacks a good way of figuring out what's the actual maximum
payload size that can be sent, for now we patch over the issue by
sending everything on a fragmented pipeline.

The exception is the unreliable messages. The fragmentation pipeline
doesn't cleanly handle having multiple fragmented payloads in flight
(and possibly arriving out of order). The reliable pipeline stage fixes
that, but for unreliable traffic it's better not to fragment it.

* Update adapter's CHANGELOG to reflect latest fixes

* Set baselib's maximum payload size to its default

Which is 2000 bytes. We won't need to hardcode that default value when
the new NetworkSettings API lands. So having a magical number floating
like that in the code is a temporary thing.