test, refactor, and style
Changed NetworkManager.GetParent to be renamed to NetworkManager.GetRootParent to better reflect what the method does.

Updated NetworkManager.GetRootParent to automatically return the root parent without having to pass in the GameObject that the NetworkManager component is attached to.

Added the ability to override the server and client creation process for MultiInstanceHelpers.  We could expand upon what portions of MultiInstanceHelpers can be overridden, but this at least puts a framework in place and should be useful to handle construction of customized NetworkManager instances in place of the default MultiInstanceHelpers way.