fix: Prevent an exception in UTP when host disconnects
The code was invoking the disconnect transport event callback on the
manager before doing some checks on UTP's NetworkConnection. But on a
client, that callback ends up calling Shutdown on the transport. This
disposes of the driver so the checks that follow end up throwing an
exception. The fix was to reorder the checks before the callback.

Also fix the logic that checked if there was a connection failure. It
was checking if the state of the connection was Connecting, but that's
never the case after a Disconnect event (the state will always be
Disconnected).