fix: load event completed fires early (#1379)
* fix

Only allow a host to include itself in the clients completed list of a SceneEventProgress (i.e. don't include dedicate-server only).

* style

adding adjusted comments.

* test

Updated SceneLoadingAndNotifications to run as a host and a server in order to test that when running as a server the SceneEventType.LoadEventCompleted and SceneEventType.UnloadEventCompleted events are triggered without the server being added to the ClientsThatCompleted list within the SceneEvent.

* update changelog

Updating the changelog file with this fix information.