fix: Rename 'Send Queue Batch Size' field in adapter
Its actual meaning is to be the maximum payload size. The name is not
entirely wrong, since this also happens to be the batch size for the
send queue, but the field would still have a utility if we removed
batching from the adapter, so 'Max Payload Size' is a more accurate
name.