fix: NetworkBehaviour spam generated when LogLevel is set to developer (#1631)
* fix

fixing issue with spam when setting NetworkManager.LogLevel = LogLevel.Developer.

* test

This test verifies the fix.  When the LogLevel is set to developer, log warnings should be generated when there is no NetworkObject assigned to a NetworkBehaviour and should not be generated when there is a valid NetworkObject.

* Update CHANGELOG.md

* style and update

Added notes about the ShutdownInProgress check as well reverted the NetworkLog.CurrentLogLevel requirement to generate this message back to its original <= value so this warning will be logged by default if the proper conditions to generate the message exist.