fix
Fixing issues with the NetworkPrefabPoolAdditive class that is used in several tests including the NetworkSceneManagerTests.  The issues fixed were related to object references not being around (i.e. NetworkManager) as well as came up with a better way to detect if the coroutine was running in order to determine if it needed to be stopped.