fix: NetworkSceneManager NetworkObjects in DDOL list can no longer exist (#1633)
* fix

This fixes an edge case scenario issue where NetworkObjects added to the DDOL list (populating sometimes but more frequently migrating back into newly loaded scene) can no longer exist.  This could be an artifact from some of the updates to the snapshot system.

* test fix

This fixes the same edge case scenario issue in the NetworkPrefabPool code base where pooled NetworkObjects could potentially no longer exist prior to the client being able to process them.

* Update CHANGELOG.md

* fix

Seeing if this fixes the repeated crashes on mac.

* Revert "fix"

This reverts commit fcc4bf2e2253824981cb0c8e870b94bddf1f2ddb.

* test

This validates the fix to check for valid GameObjects in the NetworkSceneManager when moving NetworkObjects into or from the DDOL.

* style

white space addition.

* Update CHANGELOG.md

updating the changelog entry for more clarity of what was fixed.