refactor
Added the ConditionalPredicateBase implementation of the IConditionalPredicate and implemented an overloaded version of WaitForConditionOrTimeOut.
Changed the original WaitForConditionOrTimeOut to just return true as all use cases of these UnityTest attribute decorated methods had access to the values and were relatively simple conditions.

I used the NetworkVariableTests as an example of how to use the a ConditionalPredicateBase derived class to simplify the scenario where you might be re-using the same code or conditional checks in many integration tests.