Fix: Server was able to be mapped as both ID 0 and ID 1
This meant the first client to connect could sometimes see messages from the server as coming from itself, and could successfully send to its own LocalClientId (even though the message would go to the server). Other clients would be able to send and receive from id 1, but that would also be the server.