update
Created a more global way to identify spawned NetworkObjects within integration tests.
During this time, I discovered some minor issues with tracking the player objects and resolved that.
Also discovered that if you add a NetworkBehaviour to the player object and try to run NetworkObjectDestroyTests.TestNetworkObjectClientDestroy it will throw an exception because the NetworkBehaviourUpdater continues to try to run the NetworkBehaviour even though it no longer exists (i.e. the NetworkObject.ChildNetworkBehaviours will still contain NetworkBehaviours that no longer exist).
While the issue still exists, I made adjustments to the ObjectNameIdentifier that takes this into consideration until it is determined to be a bug or expected.