fix: NetcodeIntegrationTest destroy GameObjects not NetworkObjects during shutdown sequence (#1756)
* fix

Destroy the GameObject and not the NetworkObject during DestroySceneNetworkObjects

* style

Added comment.

* fix

Some scenarios the NetworkObject and/or the GameObject can be null.
This update takes that into account.

* refactor fix

This is a cleaner and more accurate fix for the issue where a parent is destroyed before its child/children.  Added additional comments to explain why.