update
This is a WIP version.
This file needs to be broken up into several different individual tests with accompanying individual files.
There are issues with the way this test was halfway converted and for now the first thing is to do a full conversion which includes breaking this mong-test into individual tests.