test and update
adding a test to validate that checking for a NetworkObject, removing the NetworkObject, and notifying the user that a NetworkManager cannot exist if a NetworkObject resides on the GameObject.  This also now checks children containing a NetworkObject.  It does not need to check for a parent as this is checked after to invoking NetworkManagerCheckForParent.