fix: notify networkmanager and networkobject not allowed (#1777)
* fix

notify the user when they try to add a NetworkObject to a GameObject that already has a NetworkManager assigned to it.

* fix

extending the INetworkManagerHelper interface.

* update and style

Renamed the method being invoked to NotifyUserNetworkObjectRemoved in order to be more explicit in what was happening.
Updated comments

* Update CHANGELOG.md

* test and update

adding a test to validate that checking for a NetworkObject, removing the NetworkObject, and notifying the user that a NetworkManager cannot exist if a NetworkObject resides on the GameObject.  This also now checks children containing a NetworkObject.  It does not need to check for a parent as this is checked after to invoking NetworkManagerCheckForParent.

* update

making a method to return the error message.
adjusting the error message to include children in the description.
adding comments for clarity.

* update

Adding a validation check to assure the NetworkObject has been removed.

* style

whitespace fix