fix
This fixes the issue with a player not being added to the client's owned objects list.  This also fixes the issue where dynamically spawned NetworkObjects were not getting added to the client's owned object list.
This fix migrates the table into NetworkSpawnManager, removes the OwnedObjects list from NetworkClient (replaces it with a method), and provides automatic assignment and de-assignment of ownership to the NetworkSpawnManager's OwnershipToObjectsTable.