fix: key not found exception when client never connects (#1821)
* fix

This fixes the issue where if a client fails to connect after a specific period of time it will throw an exception because it never populated the m_TransportIdToClientIdMap.

* update

this is a better way of handling this

* test

The test for this fix where when a client cannot reach the server (i.e. times out) it would throw an exception complaining about a key not being found.  This test validates the fix for both UNet and UTP.

* Update CHANGELOG.md

* update

changing the default return to be zero when a client fails to connect.