fix
This is an editor side only fix to determine if the currently opened and active scene in the editor is not in the build list and if a NetworkManager instance exists that has scene management enabled.  If so, then it will display a dialog box informing the user that the scene should be added to the scenes in build list as well as provide the user the option to have it done for them.