fix: NetworkSceneManager notifications for scene not in build settings [MTT-2941] (#1828)
* fix

This provides users with more meaningful information if they are using something like ParrelSync or Multiplayer Virtual projects and encounter the Scene hash does not exist in the HashToBuildIndex table exception.

* fix

This is an editor side only fix to determine if the currently opened and active scene in the editor is not in the build list and if a NetworkManager instance exists that has scene management enabled.  If so, then it will display a dialog box informing the user that the scene should be added to the scenes in build list as well as provide the user the option to have it done for them.

* fix

Fixing issue for recent update to where Unity Transport lives for Unity versions v2022.0a5 or higher.

* fix

* Style

Added summary for ScenesInBuildActiveSceneCheck

* Update CHANGELOG.md

* Update CHANGELOG.md

MTT-2941
Final changelog modification for better user notification clarity.