fix
MTT-3041
This fixes the issue when using a single scene containing both a NetworkManager and one or more in-scene placed NetworkObjects where a client could disconnect but not unload the scene and reconnect only to receive a soft synchronization error.
The fix is to check for both HasValue and Value being true in order to account for already previously instantiated NetworkObjects.