update
Fixing some issues Fatih caught.
Refactoring when we have to delay for the authority check.  It is only needed when the parent changes.
Removing data size spam messages when LogLevel.Developer.
Adding scale and rotation to the test as well as verification that there is no notification when below the threshold.