fix
Fixed issue with ConnectionApprovalHandler's prefab getting destroyed when a client was disconnected.
Applied same fix to the NetcodeIntegrationTest when creating a prefab.

Also fixing issue with NetworkManager spewing spam messages if the NetworkConfig.ConnectionApproval is set but there is no ConnectionApprovalCallback assigned.  Adjusted it to make sure it is either the Server or Host that this applies to, but not a client.