refactor: Remove SIPTransport and replace it with UnityTransport (#1870)
* delete `SIPTransport` &  `SIPTransportTests`

* fix compile errors

* fix

This fixes the timing issues with the MessageOrdering.SpawnRpcDespawn integration test.
This also includes a minor update to NetcodeIntegrationTestHelpers (destroy as opposed to destroy immediate).

* fix

This fixes the issue with MultiClientConnectionApproval.ClientConnectedCallbackTest failing when using UTP as opposed to SIP.

* fix

Fixed issue with ConnectionApprovalHandler's prefab getting destroyed when a client was disconnected.
Applied same fix to the NetcodeIntegrationTest when creating a prefab.

Also fixing issue with NetworkManager spewing spam messages if the NetworkConfig.ConnectionApproval is set but there is no ConnectionApprovalCallback assigned.  Adjusted it to make sure it is either the Server or Host that this applies to, but not a client.

* fix

This fixes the ConnectionApprovalMismatchTest issue that caused it to fail with UTP.
Switching over to client notifications (i.e. the server detects the mismatch and disconnects the clients, we want to make sure the clients received the disconnection notification).

* fix

Global fix for teardown and NetworkObject destroying....make sure the server is the NetworkManagerOwner before destroying it.

* fix

Some NetcodeIntegrationTest adjustments to handle the switch over to UTP.

* fix

Fixes issues with NetworkObjectParentingTests (increased wait time between changes to 1 full Network Tick).
Also the recent NetcodeIntegrationTest updates helped fix some of the issues with NetworkObjectParentingTests.

* fix

Fixes the issue with NetworkSpawnManagerTests.TestConnectAndDisconnect.

* Fix - NetworkSceneManager

Switching to UnityTransport exposed a potential bug where the server disconnects a client while the client is still processing a SceneEventType.Synchronize message which by the time the client processes this it has already sent the SceneEventType.SynchronizeComplete message back to the server that *could* attempt to send a resynchronization message where the MessageSystem will crash because the m_SendQueues entry for the now disconnected client no longer exists.

* Fix - NetcodeIntegrationTest

Added some additional UnityTransport settings for payload size, send queue size, and reducing the connection attempts and time out to connect per attempt.

Added the ability to specify whether you want the server or clients to be instantiated first (see comments under NetcodeIntegrationTest.m_CreateServerFirst for further explanation).

* Fix - ClientOnlyConnectionTests

Removing UNet and just using UTP.
Reducing the timeout period so it doesn't take over 60 seconds to timeout.

* Fix - DisconnectTests

Fixes timing issues with this test that is causing it to not passing when using UTP.

* fix - NetworkObjectOwnershipTests

Fixed timing issues exposed when using UTP.

* fix - NetworkSceneManagerEventDataPoolTest

This fixes an issue when I originally switched it to be the server as the default NetworkManager to be instantiated when using NetcodeIntegrationTestHelper.Create.  This sets the, recently updated NetcodeIntegrationTest  m_CreateServerFirst, clients to instantiate before the server (needed for in-scene objects)

* fix - NetworkSceneManagerSeneVerification

This test requires the clients to be instantiated before the server.

* style

* style

* update MTT-3016

MTT-3016
Changelog entries.

* Update CHANGELOG.md

* Update CHANGELOG.md

* fix - tools

This resolves the issues with tools integration tests and switching over to UnityTransport.

* Update com.unity.netcode.gameobjects/CHANGELOG.md

Co-authored-by: Fatih Mar <mfatihmar@gmail.com>

* Update com.unity.netcode.gameobjects/CHANGELOG.md

Co-authored-by: Fatih Mar <mfatihmar@gmail.com>

* fix

The following fixes some edge-case timing related issues that are being exposed in Yamato where it might take longer than 1 network tick depending upon the virtual machine's current load.  Primarily, the majority of the fixes are handled by just waiting for the condition or timing out.

* fix - console - ps4

the time out was never yielding and causing issues on PS4/

* revert

Removing some code that didn't need to make it into this PR.

Co-authored-by: Fatih Mar <mfatihmar@gmail.com>