fix metrics tests
SceneEventTest:
This fixes an edge case issue with SceneEventTest when listening for the received load messages where in the editor it would almost always have just received the resynchronize message when it started waiting for the load message and the found metric would be incorrect.  However, when running as a stand alone testrunner build it would always have already processed the resynchronize message (on Windows) and so the resolution was to see if the message was a resynchronize message and if so create a new wait metric and listen for the expected Load message.

PacketLossMetricsTests:
I disabled scene management as the simulated packet dropping seems to be impacted possibly by the additional resynchronization message.  Also added wait for condition checks in place of the WaitForMetricsReceived as it appears 60 frames can fail when running on Windows as a stand alone testrunner build.  This is a suggested fix, but will get MTT-Tools to determine how they want to handle this (or if this is fine).