fix metric test
Fixed TrackTotalNumberOfBytesSent tracking the number of bytes received and not sent.

Fixed an edge case scenario where the server sends a resynchronization message when in the middle of tracking the number of bytes sent which cause the test to fail.  Disabled scene management to avoid this scenario for this test since it is really only focusing on testing that sent/received byte metrics are working.